requirements.txt
Pillow==10.0.0
tkinterdnd2==0.3.0
watchdog==3.0.0
opencv-python==4.8.0.76
